devices/profile/requirements/requirements-P2.txt
opencv-python-headless==4.10.0.84
numpy==1.26.4
paho-mqtt==2.1.0
prometheus-client==0.20.0